// web/src/lib/api/schema.d.ts
export interface components {
    schemas: {
// ...
        YearSpend: {
            /** Format: int32 */
            year: number;
            /** Format: int32 */
            awards: number;
            /** Format: double */
            amountEur: number;
            /** Format: double */
            approvedEur: number;
            /** Format: double */
            paidEur: number;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}